+++ src/main/java/br/com/schumaker/force/framework/ioc/managed/ManagedConfiguration.java
@@ -1,7 +1,7 @@
-package br.com.schumaker.octopus.framework.ioc.managed;
+package br.com.schumaker.force.framework.ioc.managed;
 
-import br.com.schumaker.octopus.framework.ioc.reflection.ClassReflection;
-import br.com.schumaker.octopus.framework.ioc.reflection.MethodReflection;
+import br.com.schumaker.force.framework.ioc.reflection.ClassReflection;
+import br.com.schumaker.force.framework.ioc.reflection.MethodReflection;
 
 /**
  * The ManagedConfiguration class represents a managed configuration within the IoC container.
