+++ src/main/java/br/com/schumaker/force/framework/ioc/managed/ManagedService.java
@@ -1,6 +1,6 @@
-package br.com.schumaker.octopus.framework.ioc.managed;
+package br.com.schumaker.force.framework.ioc.managed;
 
-import br.com.schumaker.octopus.framework.ioc.reflection.ClassReflection;
+import br.com.schumaker.force.framework.ioc.reflection.ClassReflection;
 
 /**
  * The ManagedService class represents a managed service within the IoC container.
