+++ src/main/java/br/com/schumaker/force/framework/ioc/reflection/validation/MaxValidation.java
@@ -1,9 +1,9 @@
-package br.com.schumaker.octopus.framework.ioc.reflection.validation;
+package br.com.schumaker.force.framework.ioc.reflection.validation;
 
-import br.com.schumaker.octopus.framework.annotations.validations.Max;
-import br.com.schumaker.octopus.framework.annotations.validations.Validate;
-import br.com.schumaker.octopus.framework.exception.OctopusException;
-import br.com.schumaker.octopus.framework.web.http.Http;
+import br.com.schumaker.force.framework.annotations.validations.Max;
+import br.com.schumaker.force.framework.annotations.validations.Validate;
+import br.com.schumaker.force.framework.exception.ForceException;
+import br.com.schumaker.force.framework.web.http.Http;
 
 import java.lang.reflect.Field;
 
@@ -25,10 +25,10 @@ public final class MaxValidation implements Validation {
      *
      * @param object the object to validate.
      * @param field the field to validate.
-     * @throws OctopusException if the field exceeds the maximum value.
+     * @throws ForceException if the field exceeds the maximum value.
      */
     @Override
-    public void validate(Object object, Field field) throws OctopusException {
+    public void validate(Object object, Field field) throws ForceException {
         try {
             field.setAccessible(true);
             double maxValue = field.getAnnotation(Max.class).value();
@@ -38,13 +38,13 @@ public final class MaxValidation implements Validation {
                 if (value > maxValue) {
                     var message = field.getAnnotation(Max.class).message();
                     if (message.equals(MAX_VALIDATION_MESSAGE)) {
-                        throw new OctopusException(String.format(message, field.getName(), maxValue), Http.HTTP_400);
+                        throw new ForceException(String.format(message, field.getName(), maxValue), Http.HTTP_400);
                     }
-                    throw new OctopusException(message);
+                    throw new ForceException(message);
                 }
             }
         } catch (IllegalAccessException e) {
-            throw new OctopusException(e.getMessage());
+            throw new ForceException(e.getMessage());
         }
     }
 }
